lamba function
AWS lambda function that performs temperature conversion and stores the results in a DynamoDB table.
The code configures the AWS resources.
The 'lambda_handler' function is the entry point for AWS lambda. It takes two parameters 'event' and 'context', which are provided by the lambda service. The code checks the input scale and performs the temperature conversion. The converted temperature is assigned to the 'mathResult' variable.
The result of the conversion along with the current time is stored in the DynamoDB table using the 'put_item' method.
Finally, it constructs a JSON response with a status code of 200 and includes a message with the converted temperature.
This code is designed to be deployed as a serverless function on AWS lambda, providing temperature conversion as a service.